Simplify sharing draft list
Instead of grouping by status, we cut it to two groups: mine and others’

This both made the dropdown easier to comprehend and also made it easier
for us to switch away from using some deprecated functions and use
`get_posts` instead.

It’s notable that we’re not showing only drafts, but also private,
future, and pending posts, too. The only difference from before is
including private posts.